Migrated from int to double on rate

+++ src/main/java/edu/ntnu/idatt2106/boco/models/Rating.java
@@ -36,7 +36,7 @@ public class Rating {
     private User user;
 
 
-    public Rating(int rate, String feedback, Rental rental, User user){
+    public Rating(double rate, String feedback, Rental rental, User user){
         this.rate = rate;
         this.feedback = feedback;
         this.rental = rental;
